Lift calculation updated
The way lift forces and moments are calculated was changed. Before, it was assumed all normal forces were concentrated in the rocket's center of pressure but this creates a distortion since the angle of attack in the nosecone is different from the angle in the fins. Now, the normal forces are calculated separately for nosecone, fins and tail.